Change: Switch from Nikola to Pelican
After looking at Nikola in more detail it became clear that its codebase
is rather dated and very difficult to understand. Sadly Pelican in
contrast was missing some features which I use (toc, teaser). Therefore
I did take a look at Pelican and it is much easier to understand and to
extend. Therefore I did implement the missing features and switched to
it finally.